Make faster by caching
The code is pretty slow. It only sells maybe 15 things at a time. This
is a problem when there is a very large backpack as it might not even
get to the end of the backpack. It is also a problem when there are a
lot of non-sellable items in the backpack (like battle pets) that slow
down the selling.

Speed up the code by caching values.