Guarantee all items are sold
The previous version of the code tried to walk from start to end of the
bags, selling each item. This worked for small bags, but for large bags
the code ran out of execution time and was terminated before all items
were sold.

Sell each bag on its own execution cycle. This should give enough time
for each bag to be fully emptied before running out of execution time.

Also, now that we are confident we can sell everything in one pass, we
can now terminate the merchant session and dismount (usually the sales
are done on the back of the Traveler's Tundra mount).